requirements.txt
# Core dependencies
feedparser>=6.0.0
beautifulsoup4>=4.9.3
requests>=2.25.1
lxml>=4.6.3
python-dateutil>=2.8.1

# Database
supabase>=2.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0

# Development
types-python-dateutil>=2.8.0
types-requests>=2.26.0
pytest>=6.2.5